Updating attributes & ignorance;
The attributes here are the Packer compile Lua code. It's going to be
considered a binary file since it's generated code and I don't care what
it is. Since I'm checking in the Packer code now, I'm making the plugin
directory something that can be checked into the Repository I also need
to make it something that I can commit now.