Adding all my Neovim updates here;
There's a few that have been important, but I'm not going to list them
out here as they're too much details in not enough space. But the big
thing here is that I'm taking my Neovim configuration with me now
wherever I go which is really nice. More on that to come in later
commits.