Updating the installation of things;
This has to be the biggest change, it's obvious here too since the rest
of the files in the repo got changed too. But the biggest thing here is
that I can now run the `./install.sh` command without worrying about
installing things with Cargo or Homebrew. This should only be done when
I want to do it, so now I have an `--init` flag that needs to be added
in order to get things to get installed via the Brewfile or the `cargo
run` commands that I've hard-coded in the script.