Updating the format;
This format is interesting to me. The `\` will make everything appear on
a single line? I have to research TOML a bit more to understand that.
But the main things here is that I've extracted the tooling around
languages installed and activated in particular directories to be on
their own line with an old desktop computer icon. I also manged to move
the command duration over to it's proper place which is near where the
success of the last command executed is. That's the left-hand side.