Adding colores-git;
This uses the commit-colors project that I really appreciate a lot. This
prints a line of text of the commit-colors you've gotten while working
on your Git branch. Then it outputs that line in comparison to the work
done in the origin's default branch. Super neat to express the work
that's been done in a visual way.