Updating my neovim setup;
This is doing a lot, but I'll explain the reasoning behind it. The main
thing that this change introduces is a custom `:ReloadConfig` command
which lets me reload all of my edited Lua files using Plenary to reload
only my custom files in scope of `configs/`, `lsp/`, and `usr/` which are
all the directories in my local `lua/` folder.aka scopes.