Removing Fugitive as my workflow's changed a lot;
I've changed, so now it's time to go back to my roots. I personally like
using a lot of the shortcuts that I have setup for Git as well. Also,
the main thing that I had issues with was the fact that the Neovim TTY
didn't support the tools I have around post- and pre- commits, so I no
longer want use Git inside of my editor.