Adding support for hyperlinks from within Neovim;
I have extra characters turned on in Neovim like spaces and carriage
returns and sometimes those return icons would get caught in the links I
would try to click with "Shift + click". Turns out, Alacritty calls
these "hints" and I just needed to add the carriage return "" icon to
the end of the regex that it uses to identify the end of a link.