Fixing sound plugin;
The sound.sh plugin wasn't working properly when I connect the audio
source to something that doesn't allow the Mac to control the audio.
This happens when using the HDMI audio out. So I'm using a little TV
icon along with an HDMI input icon to make it known that the audio can't
be controlled by the Mac for that specific output.